refactor: split output console renderers
